Add offset option to load generator

+++ core/include/process/ProcessManager.h
@@ -55,7 +55,7 @@ public:
     virtual ProcessExitEvent runProcess(std::string const &cmdLine,
                                         std::string outputFile) = 0;
 
-    virtual int64_t getNumRunningProcesses() = 0;
+    virtual size_t getNumRunningProcesses() = 0;
 
     virtual bool isShutdown() const = 0;
